+++ src/main/java/store/Receipt.java
@@ -6,12 +6,12 @@ import java.util.List;
 public class Receipt {
     private final List<Amount> amounts = new ArrayList<>();
     private boolean membershipDiscount;
-    private int totalPrice = 0;
-    private int totalBuyAmount = 0;
-    private int promotionPrice = 0;
-    private int membershipAvailablePrice = 0;
-    private int membershipPrice = 0;
-    private int payment = 0;
+    private long totalPrice = 0;
+    private long totalBuyAmount = 0;
+    private long promotionPrice = 0;
+    private long membershipAvailablePrice = 0;
+    private long membershipPrice = 0;
+    private long payment = 0;
 
     public void addAmount(Amount amount) {
         amounts.add(amount);
@@ -46,7 +46,7 @@ public class Receipt {
         amounts.forEach(amount -> promotionPrice += amount.promotionPrice());
         amounts.forEach(amount -> membershipAvailablePrice += amount.membershipAvailablePrice());
         if (membershipDiscount) {
-            membershipPrice = (int) Math.floor(membershipAvailablePrice * 0.3);
+            membershipPrice = (long) Math.floor(membershipAvailablePrice * 0.3);
         }
         if (membershipPrice > 8000) {
             membershipPrice = 8000;
